refactor!(model): reorganize threads into own module
I felt that threads were causing too much noise in the channel module,
so I moved them to a module below channel. I also reorganized the
imports in each file.
