+++ model/src/channel/thread/auto_archive_duration.rs
@@ -15,7 +15,7 @@ pub enum AutoArchiveDuration {
 }
 
 impl AutoArchiveDuration {
-    /// Retrieve the length of the duration in seconds, uesd by the API
+    /// Retrieve the length of the duration in seconds, used by the API
     ///
     /// # Examples
     ///
