+++ model/src/channel/thread/mod.rs
@@ -1,0 +1,11 @@
+mod auto_archive_duration;
+mod member;
+mod metadata;
+mod news;
+mod private;
+mod public;
+
+pub use self::{
+    auto_archive_duration::AutoArchiveDuration, member::ThreadMember, metadata::ThreadMetadata,
+    news::NewsThread, private::PrivateThread, public::PublicThread,
+};
