+++ model/src/channel/thread/public.rs
@@ -1,11 +1,12 @@
 use crate::channel::{
-    permission_overwrite::PermissionOverwrite, ChannelType, ThreadMember, ThreadMetadata,
+    thread::{ThreadMember, ThreadMetadata},
+    ChannelType,
 };
 use crate::id::{ChannelId, GuildId, MessageId, UserId};
 use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, Deserialize, Eq, Hash, PartialEq, Serialize)]
-pub struct PrivateThread {
+pub struct PublicThread {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub guild_id: Option<GuildId>,
     pub id: ChannelId,
@@ -25,26 +26,22 @@ pub struct PrivateThread {
     pub parent_id: Option<ChannelId>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub rate_limit_per_user: Option<u64>,
-    pub permission_overwrites: Vec<PermissionOverwrite>,
     pub thread_metadata: ThreadMetadata,
 }
 
 #[cfg(test)]
 mod tests {
-    use super::PrivateThread;
-    use crate::{
-        channel::{AutoArchiveDuration, ChannelType, ThreadMember, ThreadMetadata},
-        id::{ChannelId, GuildId, MessageId, UserId},
-    };
+    use super::{ChannelId, ChannelType, GuildId, MessageId, ThreadMember, ThreadMetadata, UserId};
+    use crate::channel::thread::{AutoArchiveDuration, PublicThread};
     use serde_test::Token;
 
-    #[test]
     #[allow(clippy::too_many_lines)]
-    fn test_private_thread() {
-        let value = PrivateThread {
+    #[test]
+    fn test_public_thread() {
+        let value = PublicThread {
             guild_id: Some(GuildId(2)),
             id: ChannelId(1),
-            kind: ChannelType::GuildPrivateThread,
+            kind: ChannelType::GuildPublicThread,
             last_message_id: Some(MessageId(5)),
             member_count: 7,
             member: ThreadMember {
@@ -58,7 +55,6 @@ mod tests {
             owner_id: Some(UserId(3)),
             parent_id: Some(ChannelId(4)),
             rate_limit_per_user: Some(8),
-            permission_overwrites: Vec::new(),
             thread_metadata: ThreadMetadata {
                 archived: true,
                 archiver_id: Some(UserId(9)),
@@ -72,8 +68,8 @@ mod tests {
             &value,
             &[
                 Token::Struct {
-                    name: "PrivateThread",
-                    len: 13,
+                    name: "PublicThread",
+                    len: 12,
                 },
                 Token::Str("guild_id"),
                 Token::Some,
@@ -83,7 +79,7 @@ mod tests {
                 Token::NewtypeStruct { name: "ChannelId" },
                 Token::Str("1"),
                 Token::Str("type"),
-                Token::U8(12),
+                Token::U8(11),
                 Token::Str("last_message_id"),
                 Token::Some,
                 Token::NewtypeStruct { name: "MessageId" },
@@ -123,9 +119,6 @@ mod tests {
                 Token::Str("rate_limit_per_user"),
                 Token::Some,
                 Token::U64(8),
-                Token::Str("permission_overwrites"),
-                Token::Seq { len: Some(0) },
-                Token::SeqEnd,
                 Token::Str("thread_metadata"),
                 Token::Struct {
                     name: "ThreadMetadata",
